Add patches from Fedora and make sr0 the default device
Upstream is dead, but the package is needed by udisks, so we better keep it on life-support for the time being.

Many thanks to <natrio@list.ru> for hunting down the patches and making a package in AUR that I could steal.

I also changed the default device from cdrom to sr0, as the symlink is no longer created.


git-svn-id: file:///srv/repos/svn-packages/svn@135934 eb2447ed-0c53-47e4-bac8-5bc4a241df78