rules: drop the storage group
Do not add devices to the storage group by default. This is a security risk.
The typical example is that it will let a remote user access the usb drive
of a local user.

For mounting and partitoning/formatting use udisks.

If this functionality is needed add a custom rule in /etc, or create an AUR
package.


git-svn-id: file:///srv/repos/svn-packages/svn@136369 eb2447ed-0c53-47e4-bac8-5bc4a241df78