install: remove post-install
Here's to the world's shortest grace-period!

This commit removes the creation of /dev/{console,null,zero} as it is not needed
in our standard install, is unlikely to work where it is needed, is easy to add manually
and will not break any existing installations.

Please just use an automounted devtmpfs unless you are using an initrd.


git-svn-id: file:///srv/repos/svn-packages/svn@137915 eb2447ed-0c53-47e4-bac8-5bc4a241df78