fstab: remove noexec from /tmp
Mounting /tmp as noexec is recommended by some guidelines, and it probably defends against some attacks. However, some packages rely on /tmp being mounted exec, so we do that by default.

At the end of the day the admin decides the contents of fstab, so the default is not that important.


git-svn-id: file:///srv/repos/svn-packages/svn@138031 eb2447ed-0c53-47e4-bac8-5bc4a241df78