upgpkg: filesystem 2012.6-3
add lock group

This is needed by some legacy software who wants to use /run/lock/locdev to
lock devices. That dir will be created on boot with the right group ownership.

Before this change software would put these locks in /var/lock, which at the time
had permissions 1777. Turns out that this is not sufficient for all usecases.

git-svn-id: file:///srv/repos/svn-packages/svn@162078 eb2447ed-0c53-47e4-bac8-5bc4a241df78