FS#44349 update sfdisk flag
sfdisk in 2.26 has changed meaning for '-A' flag.
In 2.25 it used to be 'Switch on the bootable flag'.
But in the new version it became
'Don't create a new partition table, but only append the specified partitions'.

Use longer version of the flag '--activate' that has the same meaning in both sfdisk versions.


git-svn-id: file:///srv/repos/svn-packages/svn@234977 eb2447ed-0c53-47e4-bac8-5bc4a241df78