Bring back vi (FS#44604)
There was objection to switching the /usr/bin/vi editor fallback of
programs like visuo to /usr/bin/nano and the idea of not providing
/usr/bin/vi at all goes against POSIX.

Considering that dbscripts lack support for pushing split packages into
more than one repository, vim-minimal cannot be made available in [core]
without maintaining two copies of the PKGBUILD, which is something the
vim maintainer would rather not do (understandably so).

So, welcome back vi; I didn't miss you one bit.

(On the bright side, vim-minimal is included on the installation image.)


git-svn-id: file:///srv/repos/svn-packages/svn@237918 eb2447ed-0c53-47e4-bac8-5bc4a241df78