Return to hook + timer
Looking at the mandb code revealed that the culprit is gdbm_reorganize,
which is supposed to be run infrequently since it causes a sync. It is
run as part of purging.

Implement a compromise where entries are added to the cache by a hook
and the caches are kept clean by a daily purge.

git-svn-id: file:///srv/repos/svn-packages/svn@271824 eb2447ed-0c53-47e4-bac8-5bc4a241df78