upgpkg: nginx 1.10.2-4 (fix logrotate regression)
Logrotate currently fails and exits unclean as the http user
lost access to the log directory to fix CVE-2016-1247.
We remove the 'su' specification for logrotate and let the root
user do the job, the 'create' option is responsible for the uid/gid
of the new files after rotation. As rotate preserves the uid/gid
this just works without behaviour change without being http.

git-svn-id: file:///srv/repos/svn-packages/svn@286910 eb2447ed-0c53-47e4-bac8-5bc4a241df78