upgpkg: ilmbase 2.2.1-2
The ilmbase pkg-config file references gl and glu, packages using pkg-config
to check for ilmbase fail if it is not installed:

$ pkg-config --exists --print-errors IlmBase
Package gl was not found in the pkg-config search path.
Perhaps you should add the directory containing `gl.pc'
to the PKG_CONFIG_PATH environment variable
Package 'gl', required by 'IlmBase', not found

So make ilmbase depend on glu.


git-svn-id: file:///srv/repos/svn-packages/svn@324864 eb2447ed-0c53-47e4-bac8-5bc4a241df78