hook.sh: Improve check_dependency and check_buildexclusive
From 9069728e4ba60df6ada86610818baff11fdabf8f Mon Sep 17 00:00:00 2001
From: Eli Schwartz <eschwartz93@gmail.com>
Date: Wed, 6 Jun 2018 12:35:42 -0400
Subject: [PATCH 1/2] simplify the hook script

Instead of copying the BUILD_DEPENDS array from a subshell that sources
the dkms.conf and uses a herestring to pass that to readarray, run the
whole function in a subshell. It needs one subshell either way, and
heredocs use a temporary file which is just totally unnecessary albeit
minimal overhead in this case.

Heredocs also invoke command substitution which strips newlines, and do
not support \0 which means readarray needs to re-split the content based
on newlines instead. Neither of these are likely to matter in this
case... but why not just avoid it, if it is free to do so?
Both of these could be solved by using < <(commands) which communicates
via a file instead of command substitution. But that still results in an
unnecessary temporary file.

Instead of copying the BUILD_EXCLUSIVE_KERNEL variable from a subshell
that sources the dkms.conf and uses a herestring to pass that to
readarray, simply use command substitution to assign the value to the
variable. The herestring is once again inefficiently opening a temporary
file, except this time it is totally unnecessary as the value in the
dkms.conf is not an array and does not need to be split into one



git-svn-id: file:///srv/repos/svn-packages/svn@337320 eb2447ed-0c53-47e4-bac8-5bc4a241df78