upgpkg: scons 3.1.0-1
Upgrading to 3.1.0 and creating a split package (scons with python2 dependencies is tracked in python2-scons, until it is dropped eventually).
Building from general source tarball by emulating a 'release'. Running tests in check() for python and python2. Separating build and packaging steps.
Patching python2 version to be installable side-by-side with python version.

git-svn-id: file:///srv/repos/svn-packages/svn@358625 eb2447ed-0c53-47e4-bac8-5bc4a241df78