Remove revert that "fixed" FS#66908
This was never the right thing to do. Cyclic dependencies existed because users
decided to hand roll their own automount units with incorrect dependencies
rather than relying on the fstab generator. Upstream changes simply exposed
these always-wrong units as being wrong (albeit not in the best of ways).
Upstream will *not* be reverting anything in the short or long term here,
so neither should Arch. What upstream *is* doing is clarifying the documentation
about automounts[0] to try to discouarge this in the future.

I'm removing this revert to make sure that the next systemd build does not
include this.

[0] https://github.com/systemd/systemd/pull/16114

git-svn-id: file:///srv/repos/svn-packages/svn@388755 eb2447ed-0c53-47e4-bac8-5bc4a241df78