upgpkg: bind 9.16.5-2
- fix FS#67306
- merge bind-tools into bind.
The orginal split was made to make the dns query clients package as small as
possible by moving the daemon out.
With the bind libraries growing, it turns out that the small daemon size (0.2M)
compared to the client/libraries (1.7M) are no more meaningful.
The packaging complexity to split client/lib and server, and with the recent
change to the documentation hierarchy, achived to convice me.



git-svn-id: file:///srv/repos/svn-packages/svn@392260 eb2447ed-0c53-47e4-bac8-5bc4a241df78