1.28.0-2: Disable resolvconf support
NM 1.28 changed the way detection of the default resolv.conf handler
works; it now prefers using resolvconf if available. However, this
does not work properly if systemd-resolvconf is installed, as NM
already configures systemd-resolved over D-Bus.

So we get the choice between enabling resolvconf support (supporting
openresolv but crashing and burning if systemd-resolvconf is
installed) and disabling resolvconf support, which then allows other
network managers to use systemd-resolvconf to merge configuration for
their interfaces with NetworkManager's.

Choose the latter and disable resolvconf support.

git-svn-id: file:///srv/repos/svn-packages/svn@403084 eb2447ed-0c53-47e4-bac8-5bc4a241df78