upgpkg: usbguard 1.0.0-2: use upstream patch files and add soname deps
The accepted PGP key has changed since 0.7.5, however there is no chain
of trust and the old maintainer isn't responsive and claims the key has
been destroyed and is lost. This is a very sad situation and basically
is a re bootstrap of the initial trust. This seems to be plausible as a
very long period of time has passed and an actual project compromise
could have bubbled up in the meanwhile -- we therefor assume there has
no compromise happened.


git-svn-id: file:///srv/repos/svn-packages/svn@409860 eb2447ed-0c53-47e4-bac8-5bc4a241df78