upgpkg: pacman 6.0.0-2: switch default integrity to sha256
The upstream makepkg.conf is a pure example and expected to be
downstream opinionated. Lets take the opportunity with pacman 6.0.0
to introduce an opinionated default integrity.

The integrity by itself is not meant to be a security property but a
pure transfer check. However in our downstream the integrity can have
some limited gain (compared to nothing) for users that properly care when
those integrity values differ. This is specifically the case in highly
distributed rebuild environments like the AUR by applying the concept of
TOFU (trust on first use) -- which by default is also applied to SSH
host verification checks and serves a good purpose when handled
appropriately. Lets aid the usage of TOFU by providing a stronger
default in our downstream and give users the opportunity to easily
detect in transit modifications of packages whose packagers just
generated default integrity values.

git-svn-id: file:///srv/repos/svn-packages/svn@415800 eb2447ed-0c53-47e4-bac8-5bc4a241df78