glibc move from partially stripping to debug pkg
Valgrind supports debuginfod since 3.17.0. But it requires
the debug symbols to be available on check(). This commit
implements a workaround to only run the check function when
the debug symbols are present. If not, a warning is shown.
Also the process of building is documented on the top of
the PKGBUILD.

git-svn-id: file:///srv/repos/svn-packages/svn@445405 eb2447ed-0c53-47e4-bac8-5bc4a241df78