upgpkg: linux-hardened 5.18.5.hardened1-2: ALSA: memalloc: Drop x86-specific hack for WC allocations
The recent report for a crash on Haswell machines implied that the
x86-specific (rather hackish) implementation for write-cache memory
buffer allocation in ALSA core is buggy with the recent kernel in some
corner cases.  This patch drops the x86-specific implementation and
uses the standard dma_alloc_wc() & co generically for avoiding the bug
and also for simplification.

git-svn-id: file:///srv/repos/svn-packages/svn@449145 eb2447ed-0c53-47e4-bac8-5bc4a241df78