upgpkg: cryptsetup 2.5.0-4: copy ossl-modules/legacy.so (FS#76440)
The OpenSSL legacy provider is required for unlocking volumes that use
the Whirlpool hash function. Cryptsetup loads automatically, though we
ought to provide ossl-modules/legacy.so for that operation to succeed.

Ideally we would try to detect when the legacy provider is for certain
needed before copying over legacy.so. The detection logic, however, is
not trivial. So for now, just eat the 32K and copy it unconditionally.


git-svn-id: file:///srv/repos/svn-packages/svn@460857 eb2447ed-0c53-47e4-bac8-5bc4a241df78