fix detection of linker scripts by ldconfig
ldconfig checks files for either `GROUP` or `GNU ld script` in its header. If
both are not present and the file isn't an ELF file either, it bails out with
"... is not an ELF file - it has the wrong magic bytes at the start".
This commit adds the missing headers.

fixes FS#77365


git-svn-id: file:///srv/repos/svn-packages/svn@468145 eb2447ed-0c53-47e4-bac8-5bc4a241df78