fix cds being older than modules archive
Until JDK 20 there was no output if the global cds (class data sharing) file
classes.jsa didn't match the shipped modules file. Since we shipped
the JRE binary with the JDK modules, the warning occured with the recent release.
This commit changes the packaging behavior to replace the JRE files with the JDK
one.

related to FS#78279


git-svn-id: file:///srv/repos/svn-packages/svn@475167 eb2447ed-0c53-47e4-bac8-5bc4a241df78